+++ src/config/fields.ts
@@ -19,6 +19,7 @@ export const matchType = {
   color: /^#([A-F0-9]{6}|[A-F0-9]{3})$/i,
   url: /((([A-Z]{3,9}:(?:\/\/)?)(?:[-;:&=\+\$,\w]+@)?[A-Z0-9.-]+|(?:www.|[-;:&=\+\$,\w]+@)[A-Z0-9.-]+)((?:\/[\+~%\/.\w-_]*)?\??(?:[-\+=&;%@.\w_]*)#?(?:[\w]*))?)/i,
   email: /.+@.+\..+/,
+  slug: /^[a-z0-9_]+$/,
 }
 
 export const validatePassword = (password: string): true | string => {